Ensure 'sudo' hint is printed when rmest is run without root
In the previous change this was fixed, however, the hint printing
was embedded in lib.rs which would be inherited by other programs
that may embed rmesg.

This change gives it a proper enum in
RMesgError::OperationNotPermitted and allows callers to know when
an operation has failed due to permissions, allowing
every consumer to implement their appropriate logic to surface
that error.